refactor: CustomException 패키지명 수정
