+++ src/main/java/gift/exception/customException/CustomException.java
@@ -1,4 +1,4 @@
-package gift.exception.CustomException;
+package gift.exception.customException;
 
 import gift.exception.ErrorCode;
