+++ src/main/java/gift/exception/customException/DuplicateEmailException.java
@@ -1,12 +1,12 @@
-package gift.exception.CustomException;
+package gift.exception.customException;
 
 import gift.exception.ErrorCode;
 import org.springframework.core.MethodParameter;
 import org.springframework.validation.BindingResult;
 
-public class EmailNotFoundException extends CustomArgumentNotValidException {
+public class DuplicateEmailException extends CustomArgumentNotValidException {
 
-    public EmailNotFoundException(MethodParameter parameter,
+    public DuplicateEmailException(MethodParameter parameter,
         BindingResult bindingResult, ErrorCode errorCode) {
         super(parameter, bindingResult, errorCode);
     }
