+++ src/main/java/gift/exception/customException/EmailNotFoundException.java
@@ -1,12 +1,12 @@
-package gift.exception.CustomException;
+package gift.exception.customException;
 
 import gift.exception.ErrorCode;
 import org.springframework.core.MethodParameter;
 import org.springframework.validation.BindingResult;
 
-public class PassWordMissMatchException extends CustomArgumentNotValidException {
+public class EmailNotFoundException extends CustomArgumentNotValidException {
 
-    public PassWordMissMatchException(MethodParameter parameter,
+    public EmailNotFoundException(MethodParameter parameter,
         BindingResult bindingResult, ErrorCode errorCode) {
         super(parameter, bindingResult, errorCode);
     }
