+++ src/main/java/gift/exception/customException/PassWordMissMatchException.java
@@ -1,12 +1,12 @@
-package gift.exception.CustomException;
+package gift.exception.customException;
 
 import gift.exception.ErrorCode;
 import org.springframework.core.MethodParameter;
 import org.springframework.validation.BindingResult;
 
-public class DuplicateEmailException extends CustomArgumentNotValidException {
+public class PassWordMissMatchException extends CustomArgumentNotValidException {
 
-    public DuplicateEmailException(MethodParameter parameter,
+    public PassWordMissMatchException(MethodParameter parameter,
         BindingResult bindingResult, ErrorCode errorCode) {
         super(parameter, bindingResult, errorCode);
     }
