+++ src/main/java/gift/service/ItemService.java
@@ -1,7 +1,7 @@
 package gift.service;
 
-import gift.exception.CustomException.CategoryNotFoundException;
-import gift.exception.CustomException.ItemNotFoundException;
+import gift.exception.customException.CategoryNotFoundException;
+import gift.exception.customException.ItemNotFoundException;
 import gift.exception.ErrorCode;
 import gift.model.categories.Category;
 import gift.model.item.Item;
