+++ src/main/java/gift/service/WishListService.java
@@ -1,7 +1,7 @@
 package gift.service;
 
-import gift.exception.CustomException.ItemNotFoundException;
-import gift.exception.CustomException.UserNotFoundException;
+import gift.exception.customException.ItemNotFoundException;
+import gift.exception.customException.UserNotFoundException;
 import gift.exception.ErrorCode;
 import gift.model.item.Item;
 import gift.model.user.User;
